fastapi/requirements.txt
#fastapi
#python-jose
#passlib
#bcrypt
#pymongo
#python-multipart
#uvicorn
#python-dotenv


#FASTAPI 
fastapi
pydantic
typing
pymysql
python-dotenv
uvicorn